Added a suffix to the game data class files
