+++ src/backend/backend.pro
@@ -1,7 +1,7 @@
 TEMPLATE = lib
 
 QT += qml quick gamepad sql
-CONFIG += c++11 staticlib warn_on exceptions_off object_parallel_to_source
+CONFIG += c++11 staticlib warn_on exceptions_off
 android: QT += androidextras
