+++ src/backend/modeldata/gaming/GameAssetsData.cpp
@@ -15,7 +15,7 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
 
-#include "GameAssets.h"
+#include "GameAssetsData.h"
 
 #include <QUrl>
