+++ src/backend/providers/android_apps/AndroidAppsMetadata.cpp
@@ -18,8 +18,8 @@
 #include "AndroidAppsMetadata.h"
 
 #include "LocaleUtils.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 #include "providers/JsonCacheUtils.h"
 #include "utils/HashMap.h"
