+++ src/backend/providers/android_apps/AndroidAppsProvider.cpp
@@ -17,8 +17,8 @@
 
 #include "AndroidAppsProvider.h"
 
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 
 #include <QFileInfo>
 #include <QtAndroidExtras/QAndroidJniEnvironment>
