+++ src/backend/providers/gog/GogGamelist.cpp
@@ -20,8 +20,8 @@
 #include "GogCommon.h"
 #include "LocaleUtils.h"
 #include "Paths.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 #include "utils/MoveOnly.h"
 
 #include <QDebug>
