+++ src/backend/providers/pegasus/PegasusCollections.cpp
@@ -22,8 +22,8 @@
 #include "Paths.h"
 #include "PegasusAssets.h"
 #include "PegasusCommon.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 #include "utils/PathCheck.h"
 
 #include <QDebug>
