+++ src/backend/providers/pegasus/PegasusCommon.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "modeldata/gaming/GameAssets.h"
+#include "modeldata/gaming/GameAssetsData.h"
 
 #include <QString>
 #include <QStringList>
