+++ src/backend/providers/pegasus/PegasusMetadata.cpp
@@ -22,7 +22,7 @@
 #include "Paths.h"
 #include "PegasusAssets.h"
 #include "PegasusCommon.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/GameData.h"
 #include "utils/PathCheck.h"
 
 #include <QDebug>
