+++ src/backend/providers/steam/SteamMetadata.cpp
@@ -19,8 +19,8 @@
 
 #include "LocaleUtils.h"
 #include "Paths.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 #include "providers/JsonCacheUtils.h"
 
 #include <QDebug>
