+++ tests/backend/providers/pegasus/test_PegasusProvider.cpp
@@ -18,8 +18,8 @@
 #include <QtTest/QtTest>
 
 #include "providers/pegasus/PegasusProvider.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 #include "utils/HashMap.h"
 
 #include <QString>
