+++ tests/benchmarks/pegasus_provider/bench_PegasusProvider.cpp
@@ -18,8 +18,8 @@
 #include <QtTest/QtTest>
 
 #include "providers/pegasus/PegasusProvider.h"
-#include "modeldata/gaming/Collection.h"
-#include "modeldata/gaming/Game.h"
+#include "modeldata/gaming/CollectionData.h"
+#include "modeldata/gaming/GameData.h"
 
 #include <QString>
